Pull tests out into seperate files to help, add test for Max.
#74
